+++ tasks/test.sh
@@ -1,8 +1,3 @@
 #!/bin/bash
 
-ERRORS=0
-# pylint --rcfile="test/.pylintrc" test || ERRORS=1
-
-if [ "$ERRORS" -eq 0 ]; then
-	python3.5 test/testrunner.py
-fi
+python3.5 test/testrunner.py